+++ stroom-legacy/stroom-legacy-db-migration/src/main/java/stroom/legacy/db/migration/V07_00_00_006__XmlSchema.java
@@ -16,6 +16,7 @@
 
 package stroom.legacy.db.migration;
 
+import stroom.db.util.DbUtil;
 import stroom.docstore.impl.Serialiser2FactoryImpl;
 import stroom.pipeline.xmlschema.XmlSchemaSerialiser;
 import stroom.xmlschema.shared.XmlSchemaDoc;
@@ -54,9 +55,9 @@ public class V07_00_00_006__XmlSchema extends BaseJavaMigration {
                         "FROM OLD_XML_SCHEMA")) {
             try (final ResultSet resultSet = preparedStatement.executeQuery()) {
                 while (resultSet.next()) {
-                    final Long crtMs = (Long) resultSet.getObject(1);
+                    final Long crtMs = DbUtil.getLong(resultSet, 1);
                     final String crtUser = resultSet.getString(2);
-                    final Long updMs = (Long) resultSet.getObject(3);
+                    final Long updMs = DbUtil.getLong(resultSet, 3);
                     final String updUser = resultSet.getString(4);
                     final String name = resultSet.getString(5);
                     final String uuid = resultSet.getString(6);