+++ stroom-legacy/stroom-legacy-db-migration/src/main/java/stroom/legacy/db/migration/V07_00_00_008__Dashboard.java
@@ -18,6 +18,7 @@ package stroom.legacy.db.migration;
 
 import stroom.dashboard.impl.DashboardSerialiser;
 import stroom.dashboard.shared.DashboardDoc;
+import stroom.db.util.DbUtil;
 import stroom.docstore.impl.Serialiser2FactoryImpl;
 import stroom.legacy.impex_6_1.MappingUtil;
 import stroom.legacy.model_6_1.Dashboard;
@@ -53,9 +54,9 @@ public class V07_00_00_008__Dashboard extends BaseJavaMigration {
                         "FROM OLD_DASH")) {
             try (final ResultSet resultSet = preparedStatement.executeQuery()) {
                 while (resultSet.next()) {
-                    final Long crtMs = (Long) resultSet.getObject(1);
+                    final Long crtMs = DbUtil.getLong(resultSet, 1);
                     final String crtUser = resultSet.getString(2);
-                    final Long updMs = (Long) resultSet.getObject(3);
+                    final Long updMs = DbUtil.getLong(resultSet, 3);
                     final String updUser = resultSet.getString(4);
                     final String name = resultSet.getString(5);
                     final String uuid = resultSet.getString(6);