+++ stroom-legacy/stroom-legacy-db-migration/src/main/java/stroom/legacy/db/migration/V07_00_00_016__Feed.java
@@ -58,9 +58,9 @@ public class V07_00_00_016__Feed extends BaseJavaMigration {
                         "LEFT OUTER JOIN OLD_STRM_TP st ON (st.ID = f.FK_STRM_TP_ID)")) {
             try (final ResultSet resultSet = preparedStatement.executeQuery()) {
                 while (resultSet.next()) {
-                    final Long crtMs = (Long) resultSet.getObject(1);
+                    final Long crtMs = DbUtil.getLong(resultSet, 1);
                     final String crtUser = resultSet.getString(2);
-                    final Long updMs = (Long) resultSet.getObject(3);
+                    final Long updMs = DbUtil.getLong(resultSet, 3);
                     final String updUser = resultSet.getString(4);
                     final String name = resultSet.getString(5);
                     final String uuid = resultSet.getString(6);